requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
iniconfig==1.1.1
jws==0.1.3
oauth2client==4.1.3
packaging==21.0
pluggy==1.0.0
protobuf==3.18.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.4
pyparsing==2.4.7
Pyrebase4==4.5.0
pytest==6.2.5
python-jwt==2.0.1
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
toml==0.10.2
urllib3==1.26.7
yapf==0.31.0
